8169725: cannot use TIFFField(TIFFTag tag, int value) for TIFF_LONG values greater than Integer.MAX_VALUE
Change constructor TIFFField(TIFFTag,int) to TIFFField(TIFFTag,long).

+++ jdk/src/java.desktop/share/classes/com/sun/imageio/plugins/tiff/TIFFImageMetadata.java
@@ -87,7 +87,7 @@ public class TIFFImageMetadata extends IIOMetadata {
         rootIFD.initialize(stream, true, ignoreUnknownFields);
     }
 
-    public void addShortOrLongField(int tagNumber, int value) {
+    public void addShortOrLongField(int tagNumber, long value) {
         TIFFField field = new TIFFField(rootIFD.getTag(tagNumber), value);
         rootIFD.addTIFFField(field);
     }
